+++ package.json
@@ -34,6 +34,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.1",
+    "react-icons": "^4.11.0",
     "swr": "^2.0.4",
     "typescript": "4.9.4",
     "uuid": "^9.0.1",
